feat: move FeedFilters to commons (batch 7)
Move the following FeedFilters from amethyst to commons module:
- GeoHashFeedFilter
- HashtagFeedFilter
- CommunityFeedFilter + CommunityModerationFeedFilter
- UserProfileConversationsFeedFilter
- UserProfileNewThreadFeedFilter

Interface changes to support this:
- ICacheProvider: add notes/addressables collection-level ICacheOperations
- IAccount: add liveHiddenUsers() for content filtering

Amethyst-side files replaced with typealiases. VMs remain as thin
Android wrappers passing LocalCache to the commons filters.
