+++ amethyst/src/main/java/com/vitorpamplona/amethyst/ui/screen/loggedIn/communities/dal/CommunityModerationFeedFilter.kt
@@ -20,107 +20,5 @@
  */
 package com.vitorpamplona.amethyst.ui.screen.loggedIn.communities.dal
 
-import com.vitorpamplona.amethyst.model.Account
-import com.vitorpamplona.amethyst.model.AddressableNote
-import com.vitorpamplona.amethyst.model.LocalCache
-import com.vitorpamplona.amethyst.model.Note
-import com.vitorpamplona.amethyst.ui.dal.AdditiveFeedFilter
-import com.vitorpamplona.amethyst.ui.dal.DefaultFeedOrder
-import com.vitorpamplona.quartz.nip10Notes.TextNoteEvent
-import com.vitorpamplona.quartz.nip18Reposts.GenericRepostEvent
-import com.vitorpamplona.quartz.nip18Reposts.RepostEvent
-import com.vitorpamplona.quartz.nip22Comments.CommentEvent
-import com.vitorpamplona.quartz.nip72ModCommunities.approval.CommunityPostApprovalEvent
-import com.vitorpamplona.quartz.nip72ModCommunities.definition.CommunityDefinitionEvent
-import com.vitorpamplona.quartz.nip72ModCommunities.isForCommunity
-
-class CommunityModerationFeedFilter(
-    val communityDefNote: AddressableNote,
-    val account: Account,
-) : AdditiveFeedFilter<Note>() {
-    val approvedFilter = CommunityFeedFilter(communityDefNote, account)
-
-    val communityDefEvent = communityDefNote.event as? CommunityDefinitionEvent
-    val moderators = communityDefEvent?.moderatorKeys()?.toSet() ?: emptySet()
-
-    override fun feedKey(): String = account.userProfile().pubkeyHex + "-" + communityDefNote.idHex
-
-    override fun feed(): List<Note> {
-        if (communityDefEvent == null) return emptyList()
-
-        val result =
-            LocalCache.notes.mapFlattenIntoSet { _, it ->
-                filterMap(it)
-            }
-
-        return sort(result)
-    }
-
-    override fun updateListWith(
-        oldList: List<Note>,
-        newItems: Set<Note>,
-    ): List<Note> {
-        val approved = approvedFilter.applyFilter(newItems)
-
-        val newOldList =
-            if (approved.isNotEmpty()) {
-                oldList - approved
-            } else {
-                oldList
-            }
-
-        val newItemsToBeAdded = applyFilter(newItems)
-        return if (newItemsToBeAdded.isNotEmpty()) {
-            val newList = newOldList.toSet() + newItemsToBeAdded
-            sort(newList).take(limit())
-        } else {
-            newOldList
-        }
-    }
-
-    override fun applyFilter(newItems: Set<Note>): Set<Note> = innerApplyFilter(newItems)
-
-    private fun innerApplyFilter(collection: Collection<Note>): Set<Note> {
-        if (communityDefEvent == null) return emptySet()
-
-        return collection
-            .mapNotNull {
-                filterMap(it)
-            }.flatten()
-            .toSet()
-    }
-
-    private fun wasApprovedByCommunity(note: Note): Boolean =
-        note.boosts.any {
-            val approvalEvent = it.event
-            (approvalEvent is CommunityPostApprovalEvent || approvalEvent is RepostEvent || approvalEvent is GenericRepostEvent) &&
-                approvalEvent.pubKey in moderators &&
-                approvalEvent.isForCommunity(this.communityDefNote.idHex)
-        }
-
-    private fun filterMap(note: Note): List<Note>? {
-        val noteEvent = note.event ?: return null
-
-        return if (
-            (noteEvent is TextNoteEvent || noteEvent is CommentEvent || noteEvent is CommunityPostApprovalEvent) &&
-            noteEvent.isForCommunity(this.communityDefNote.idHex) &&
-            noteEvent.pubKey !in moderators
-        ) {
-            if (noteEvent is CommunityPostApprovalEvent) {
-                note.replyTo?.filter {
-                    it.isNewThread() && !wasApprovedByCommunity(it)
-                }
-            } else {
-                if (note.isNewThread() && !wasApprovedByCommunity(note)) {
-                    listOf(note)
-                } else {
-                    null
-                }
-            }
-        } else {
-            null
-        }
-    }
-
-    override fun sort(items: Set<Note>): List<Note> = items.sortedWith(DefaultFeedOrder)
-}
+// Re-export from commons for backwards compatibility
+typealias CommunityModerationFeedFilter = com.vitorpamplona.amethyst.commons.ui.feeds.community.CommunityModerationFeedFilter
