+++ amethyst/src/main/java/com/vitorpamplona/amethyst/ui/screen/loggedIn/geohash/dal/GeoHashFeedFilter.kt
@@ -20,72 +20,5 @@
  */
 package com.vitorpamplona.amethyst.ui.screen.loggedIn.geohash.dal
 
-import com.vitorpamplona.amethyst.model.Account
-import com.vitorpamplona.amethyst.model.LocalCache
-import com.vitorpamplona.amethyst.model.Note
-import com.vitorpamplona.amethyst.ui.dal.AdditiveFeedFilter
-import com.vitorpamplona.amethyst.ui.dal.DefaultFeedOrder
-import com.vitorpamplona.quartz.experimental.audio.header.AudioHeaderEvent
-import com.vitorpamplona.quartz.experimental.zapPolls.ZapPollEvent
-import com.vitorpamplona.quartz.nip01Core.core.Event
-import com.vitorpamplona.quartz.nip01Core.relay.normalizer.NormalizedRelayUrl
-import com.vitorpamplona.quartz.nip01Core.tags.geohash.isTaggedGeoHash
-import com.vitorpamplona.quartz.nip04Dm.messages.PrivateDmEvent
-import com.vitorpamplona.quartz.nip10Notes.TextNoteEvent
-import com.vitorpamplona.quartz.nip22Comments.CommentEvent
-import com.vitorpamplona.quartz.nip23LongContent.LongTextNoteEvent
-import com.vitorpamplona.quartz.nip28PublicChat.message.ChannelMessageEvent
-import com.vitorpamplona.quartz.nip54Wiki.WikiNoteEvent
-import com.vitorpamplona.quartz.nip73ExternalIds.location.GeohashId
-
-class GeoHashFeedFilter(
-    val tag: String,
-    val relays: Set<NormalizedRelayUrl>,
-    val account: Account,
-    val cache: LocalCache,
-) : AdditiveFeedFilter<Note>() {
-    override fun feedKey(): String = account.userProfile().pubkeyHex + "-" + tag
-
-    override fun feed(): List<Note> {
-        val notes =
-            cache.notes.filterIntoSet { _, it ->
-                acceptableEvent(it, tag)
-            }
-
-        return sort(notes)
-    }
-
-    override fun applyFilter(newItems: Set<Note>): Set<Note> = innerApplyFilter(newItems)
-
-    private fun innerApplyFilter(collection: Collection<Note>): Set<Note> = collection.filterTo(HashSet()) { acceptableEvent(it, tag) }
-
-    fun acceptableEvent(
-        it: Note,
-        geoTag: String,
-    ): Boolean =
-        (acceptableViaHashtag(it.event, geoTag) || acceptableViaScope(it.event, geoTag)) &&
-            !it.isHiddenFor(account.hiddenUsers.flow.value) &&
-            account.isAcceptable(it)
-
-    fun acceptableViaHashtag(
-        event: Event?,
-        geohash: String,
-    ): Boolean =
-        (
-            event is TextNoteEvent ||
-                event is LongTextNoteEvent ||
-                event is WikiNoteEvent ||
-                event is ChannelMessageEvent ||
-                event is PrivateDmEvent ||
-                event is ZapPollEvent ||
-                event is AudioHeaderEvent
-        ) &&
-            event.isTaggedGeoHash(geohash)
-
-    fun acceptableViaScope(
-        event: Event?,
-        geohash: String,
-    ): Boolean = event is CommentEvent && event.isTaggedScope(geohash, GeohashId::match)
-
-    override fun sort(items: Set<Note>): List<Note> = items.sortedWith(DefaultFeedOrder)
-}
+// Re-export from commons for backwards compatibility
+typealias GeoHashFeedFilter = com.vitorpamplona.amethyst.commons.ui.feeds.geohash.GeoHashFeedFilter
